+++ lib/ex_phone_number/mix/tasks/update_metadata.ex
@@ -1,46 +1,38 @@
-if Code.ensure_loaded?(Tesla) do
+if Code.ensure_loaded?(Req) do
   defmodule Mix.Tasks.UpdateMetadata do
     @moduledoc "Downloads the latest libphonenumber metadata from GitHub"
     @shortdoc "Update libphonenumber metadata"
 
     use Mix.Task
+    require Logger
+
     @raw_files_url "https://raw.githubusercontent.com/google/libphonenumber"
-    @files_to_download ["resources/PhoneNumberMetadata.xml", "resources/PhoneNumberMetadataForTesting.xml"]
+    @files_to_download [
+      "resources/PhoneNumberMetadata.xml",
+      "resources/PhoneNumberMetadataForTesting.xml"
+    ]
     @resources_directory "resources"
-
-    defmodule GitHub do
-      def latest_release(repo) do
-        Tesla.get(client(), "/repos/" <> repo <> "/releases/latest")
-      end
-
-      defp client do
-        middleware = [
-          {Tesla.Middleware.BaseUrl, "https://api.github.com"},
-          Tesla.Middleware.JSON,
-          {Tesla.Middleware.Headers, [{"User-Agent", "ex_phone_number"}]}
-        ]
-
-        Tesla.client(middleware)
-      end
-    end
 
     @impl Mix.Task
     def run(_args) do
+      Application.ensure_all_started(:req)
       latest_tag = fetch_latest_tag()
+      Logger.info("Downloading metadata #{latest_tag}")
       Enum.each(@files_to_download, &download(latest_tag, &1))
       update_readme(latest_tag)
     end
 
     defp fetch_latest_tag() do
-      {:ok, %{body: body}} = GitHub.latest_release("google/libphonenumber")
-      body["tag_name"]
+      Req.get!(req(), url: "https://api.github.com/repos/google/libphonenumber/releases/latest").body[
+        "tag_name"
+      ]
     end
 
     defp download(tag, path) do
       filename = Path.basename(path)
       local_path = Path.join([File.cwd!(), @resources_directory, filename])
       file_url = "#{@raw_files_url}/#{tag}/#{path}"
-      {:ok, %{body: body}} = Tesla.get(file_url)
+      %{status: 200, body: body} = Req.get!(req(), url: file_url)
       File.write!(local_path, body)
     end
 
@@ -50,9 +42,16 @@ if Code.ensure_loaded?(Tesla) do
       updated_content =
         readme_path
         |> File.read!()
-        |> String.replace(~r{Current metadata version: v[\d+].[\d+].[\d+]\.}, "Current metadata version: " <> tag <> ".")
+        |> String.replace(
+          ~r{Current metadata version: v\d+\.\d+\.\d+\.},
+          "Current metadata version: " <> tag <> "."
+        )
 
       File.write!(readme_path, updated_content)
+    end
+
+    defp req() do
+      Req.new(headers: [user_agent: "ex_phone_number"])
     end
   end
 end